Rename internal/artifacts.py to internal/artifact/artifact_saver.py
