+++ tests/pytest_tests/system_tests/test_core/test_cli_full.py
@@ -166,7 +166,7 @@ def test_sync_wandb_run(runner, relay_server, user, copy_asset):
     # (as we used to use a mock backend)
     # todo: create a new test asset that will contain an artifact
     with relay_server() as relay, runner.isolated_filesystem(), mock.patch(
-        "wandb.sdk.internal.artifacts.ArtifactSaver.save", return_value=None
+        "wandb.sdk.internal.artifact_saver.ArtifactSaver.save", return_value=None
     ):
         copy_asset("wandb")
 
@@ -186,7 +186,7 @@ def test_sync_wandb_run(runner, relay_server, user, copy_asset):
 
 def test_sync_wandb_run_and_tensorboard(runner, relay_server, user, copy_asset):
     with relay_server() as relay, runner.isolated_filesystem(), mock.patch(
-        "wandb.sdk.internal.artifacts.ArtifactSaver.save", return_value=None
+        "wandb.sdk.internal.artifact_saver.ArtifactSaver.save", return_value=None
     ):
         run_dir = os.path.join("wandb", "offline-run-20210216_154407-g9dvvkua")
         copy_asset("wandb")
